fixed circular import, and am running tests (downloading manga :) )
