+++ .gitignore
@@ -1,2 +1,3 @@
 **/mangaDataBackupOld.json
 **/mangaDataBackupNEW.json
+**/__pycache__/
