+++ helperfunctions.py
@@ -1,9 +1,23 @@
 import json
 import time
 from random import randint
-from mangaScrape import weebCentralBase, mangakakalotBase
-from weebcentral import rip_manga_ms
-from mangakakalot import rip_manga_mk
+from selenium import webdriver
+from selenium.webdriver import ChromeOptions
+import pathlib
+import requests
+from bs4 import BeautifulSoup
+
+options = ChromeOptions()
+options.add_argument("--headless=new")
+options.add_argument("--log-level=3")
+driver = webdriver.Chrome(options=options)
+weebCentralBase = "https://weebcentral.com"
+mangakakalotBase = "https://mangakakalot.com"
+headers = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/129.0.0.0 Safari/537.3"
+
+BASE_DLPATH = "D:/MANGA STORAGE"
+
+
 
 def insert_to_file(new_entry):
     # managed json file
@@ -97,3 +111,102 @@ def download_handler(chapter_list, manga_idx):
                 print("No more chapters to gather: Last updated at " +
                       data[manga_idx]['lastUpdated'])
                 raise Exception("No more chapters to gather")
+
+def rip_manga_ms(page, data, manga_idx):
+    driver.get(page)
+    chapter_images = []
+    time.sleep(2)
+    soup = BeautifulSoup(driver.page_source, 'html.parser')
+    chapter_folder_wrapper = soup.find("button", "col-span-4 lg:flex-1 btn btn-secondary")
+    chapter_folder = chapter_folder_wrapper.find("span").text
+    chapter_folder = clean_and_strip(chapter_folder)
+    if '-' not in chapter_folder:
+        chapter_folder = 'S0 - ' + chapter_folder
+    image_elements = soup.find_all('img', class_="maw-w-full mx-auto")
+    print(image_elements)
+    try:
+        pathlib.Path(BASE_DLPATH + "/" + data[manga_idx]['title'] +"/" + chapter_folder).mkdir(parents=True, exist_ok=False)
+    except FileExistsError:
+        return False
+    # make request for each image src :)
+    image_count = 0
+    for img in image_elements:
+        img_src = img.get('src')
+        chapter_images.append(img_src)
+    for image_url in chapter_images:
+        fileName = image_url.split('/')
+        fileName = fileName[len(fileName)-1]
+        response = requests.get(image_url, headers={'UserAgent': headers, 'referer': "https://weebcentral.com/"})
+        if (response.status_code != 200):
+            print("Error getting the current file")
+            return False
+        else:
+            with open(BASE_DLPATH + "/" + data[manga_idx]['title'] +
+                      "/" + chapter_folder + '/' + fileName, 'wb') as f:
+                noop = f.write(response.content)
+                print("Saved {}".format(BASE_DLPATH + "/" + data[manga_idx]['title'] +
+                      "/" + chapter_folder + '/' + fileName))
+        image_count += 1
+        if image_count % 10 == 0:
+            time.sleep(randint(9, 15))
+    return True
+
+def rip_manga_mk(page, data, manga_idx):
+    chapter_images = []
+    search_url = page
+    driver.get(search_url)
+    soup = BeautifulSoup(driver.page_source, 'html.parser')
+    image_container = soup.find('div', class_="container-chapter-reader")
+    chapter_folder = ""
+    # chapter folder creations
+    if "chapmanganato" in data[manga_idx]['link']:
+        chapter_selection = soup.find("select", class_="navi-change-chapter")
+        chapter_selection = chapter_selection.find("option", selected=True)
+        chapter_selection = clean_and_strip(chapter_selection.text)
+        chapter_folder += chapter_selection
+    elif "mangakakalot" in data[manga_idx]["link"]:
+        chapter_h1 = soup.find("h1", class_="current-chapter")
+        chapter_text = chapter_h1.text
+        chapter_text = chapter_text.split(": ")
+        for splt in chapter_text:
+            if "chapter" in splt or "Chapter" in splt:
+                chapter_text = clean_and_strip(splt)
+                break
+        chapter_folder += chapter_text
+    if '-' not in chapter_folder:
+        chapter_folder = "S0 - " + chapter_folder
+    try:
+        pathlib.Path(BASE_DLPATH + "/" + data[manga_idx]['title'] +
+                     "/" + chapter_folder).mkdir(parents=True, exist_ok=False)
+    except FileExistsError:
+        return False
+    # grabbing image from chapter
+    all_images = image_container.findAll('img')
+    for img in all_images:
+        if (img.has_attr('title') and img):
+            img_src = img.get('src')
+            chapter_images.append(img_src)
+    print(chapter_images)
+    image_count = 0
+    for image_url in chapter_images:
+        referer = ""
+        fileName = image_url.split('/')
+        fileName = fileName[len(fileName)-1]
+        if "chapmanganato" in data[manga_idx]['link']:
+            referer = "https://chapmanganato.to/"
+        elif "mangakakalot" in data[manga_idx]['link']:
+            referer = "https://mangakakalot.com/"
+        response = requests.get(image_url, headers={
+                                'UserAgent': headers, 'referer': referer})
+        if (response.status_code != 200):
+            print("Error getting the current file")
+            return False
+        else:
+            print("Woooo we have our images")
+            with open(BASE_DLPATH + "/" + data[manga_idx]['title'] + "/" + chapter_folder + '/' + fileName, 'wb') as f:
+                noop = f.write(response.content)
+                print("Saved {}".format(BASE_DLPATH + "/" + data[manga_idx]['title'] +
+                      "/" + chapter_folder + '/' + fileName))
+        image_count += 1
+        if image_count % 10 == 0:
+            time.sleep(randint(9, 15))
